# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Tenant Management","description":"APIs for managing tenants"},{"name":"API Key Management","description":"APIs for managing API keys"}],"paths":{"/api/v1/tenants/{id}":{"get":{"tags":["Tenant Management"],"summary":"Get tenant by ID","description":"Retrieve a specific tenant by ID","operationId":"getTenantById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TenantResponse"}}}}}},"put":{"tags":["Tenant Management"],"summary":"Update tenant","description":"Update an existing tenant","operationId":"updateTenant","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTenantRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TenantResponse"}}}}}},"delete":{"tags":["Tenant Management"],"summary":"Delete tenant","description":"Soft delete a tenant by setting status to INACTIVE","operationId":"deleteTenant","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}},"/api/v1/tenants/{id}/status":{"put":{"tags":["Tenant Management"],"summary":"Update tenant status","description":"Update tenant status (ACTIVE/INACTIVE)","operationId":"updateTenantStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantStatusUpdateRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TenantResponse"}}}}}}},"/api/v1/data-endpoints/{id}":{"get":{"tags":["data-endpoint-controller"],"operationId":"getEndpointById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataEndpoint"}}}}}},"put":{"tags":["data-endpoint-controller"],"operationId":"updateEndpoint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataEndpoint"}}}}}},"delete":{"tags":["data-endpoint-controller"],"operationId":"deleteEndpoint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}},"/api/v1/connectors/{id}":{"get":{"tags":["connector-controller"],"operationId":"getConnectorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Connector"}}}}}},"put":{"tags":["connector-controller"],"operationId":"updateConnector","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConnectorRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Connector"}}}}}},"delete":{"tags":["connector-controller"],"operationId":"deleteConnector","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}},"/api/v1/connectors/{id}/approval":{"put":{"tags":["connector-controller"],"operationId":"updateApprovalStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Connector"}}}}}}},"/api/v1/api-keys/{id}/revoke":{"put":{"tags":["API Key Management"],"summary":"Revoke API key","description":"Revoke an API key, setting its status to REVOKED","operationId":"revokeApiKey","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiKeyResponse"}}}}}}},"/api/v1/tenants":{"get":{"tags":["Tenant Management"],"summary":"Get all tenants","description":"Retrieve list of all tenants with optional status filter","operationId":"getAllTenants","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TenantResponse"}}}}}}},"post":{"tags":["Tenant Management"],"summary":"Create tenant","description":"Create a new tenant","operationId":"createTenant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TenantResponse"}}}}}}},"/api/v1/tenants/{tenantId}/api-keys":{"get":{"tags":["API Key Management"],"summary":"Get all API keys for a tenant","description":"Retrieve list of all API keys for a specific tenant","operationId":"getApiKeysByTenant","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApiKeyListItemResponse"}}}}}}},"post":{"tags":["API Key Management"],"summary":"Generate new API key","description":"Generate a new API key for a tenant. The raw key is only returned once and cannot be retrieved again.","operationId":"generateApiKey","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApiKeyRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiKeyCreationResponse"}}}}}}},"/api/v1/data-endpoints":{"get":{"tags":["data-endpoint-controller"],"operationId":"getAllEndpoints","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataEndpoint"}}}}}}},"post":{"tags":["data-endpoint-controller"],"operationId":"createEndpoint","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataEndpoint"}}}}}}},"/api/v1/data-endpoints/test":{"post":{"tags":["data-endpoint-controller"],"operationId":"testQuery","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v1/connectors":{"get":{"tags":["connector-controller"],"operationId":"getAllConnectors","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Connector"}}}}}}},"post":{"tags":["connector-controller"],"operationId":"createConnector","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConnectorRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Connector"}}}}}}},"/api/v1/connectors/test-connection":{"post":{"tags":["connector-controller"],"operationId":"testConnection","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TableMetadata"}}}}}}}},"/api/v1/data-endpoints/{id}/status":{"patch":{"tags":["data-endpoint-controller"],"operationId":"toggleStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataEndpoint"}}}}}}},"/api/v1/data/{dataEndpointId}":{"get":{"tags":["dynamic-data-controller"],"operationId":"queryDynamicData","parameters":[{"name":"dataEndpointId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/connectors/{id}/tables":{"get":{"tags":["connector-controller"],"operationId":"getTables","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TableMetadata"}}}}}}}},"/api/v1/api-keys/{id}":{"get":{"tags":["API Key Management"],"summary":"Get API key details","description":"Get details of a specific API key (raw key not included)","operationId":"getApiKeyById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiKeyResponse"}}}}}},"delete":{"tags":["API Key Management"],"summary":"Delete API key","description":"Permanently delete an API key","operationId":"deleteApiKey","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"code":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"traceId":{"type":"string"},"details":{"type":"object","additionalProperties":{"type":"object"}}}},"UpdateTenantRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}}},"TenantResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"TenantStatusUpdateRequest":{"type":"object","properties":{"status":{"type":"string"}}},"ColumnMetadata":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"dataType":{"type":"string"},"semanticType":{"type":"string","enum":["TEXT","NUMBER","BOOLEAN","DATE","DATETIME","TIME","TIMESTAMP","CURRENCY","CITY","COUNTRY","LATITUDE","LONGITUDE","ID","UUID","URL","IMAGE_URL","EMAIL","CATEGORY","STATUS","UNKNOWN"]},"description":{"type":"string"},"visibility":{"type":"string","enum":["VISIBLE","HIDDEN","TECHNICAL","EVERYWHERE","DETAIL_ONLY"]},"foreignKeyTarget":{"type":"string"},"formatting":{"type":"object","additionalProperties":{"type":"object"}},"primaryKey":{"type":"boolean"},"foreignKey":{"type":"boolean"}}},"ConnectionConfig":{"type":"object","properties":{"config":{"type":"object","additionalProperties":{"type":"object"}},"host":{"type":"string"},"port":{"type":"integer","format":"int32"},"schema":{"type":"string"},"configMap":{"type":"object","additionalProperties":{"type":"object"}},"username":{"type":"string"},"databaseName":{"type":"string"}}},"Connector":{"type":"object","properties":{"domainEvents":{"type":"array","items":{"$ref":"#/components/schemas/DomainEvent"}},"id":{"$ref":"#/components/schemas/ConnectorId"},"name":{"type":"string"},"type":{"$ref":"#/components/schemas/ConnectorType"},"config":{"$ref":"#/components/schemas/ConnectionConfig"},"tenantId":{"type":"string"},"status":{"type":"string","enum":["INIT","APPROVED","REJECTED"]},"registeredTables":{"type":"array","items":{"$ref":"#/components/schemas/TableMetadata"}},"createdAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"},"typeValue":{"type":"string"},"idValue":{"type":"string"}}},"ConnectorId":{"type":"object","properties":{"id":{"type":"string","format":"uuid"}}},"ConnectorType":{"type":"object","properties":{"type":{"type":"string"},"restApi":{"type":"boolean"},"database":{"type":"boolean"}}},"DataEndpoint":{"type":"object","properties":{"domainEvents":{"type":"array","items":{"$ref":"#/components/schemas/DomainEvent"}},"id":{"$ref":"#/components/schemas/DataEndpointId"},"connector":{"$ref":"#/components/schemas/Connector"},"name":{"type":"string"},"description":{"type":"string"},"pathAlias":{"type":"string"},"queryConfig":{"type":"string"},"fieldMaskingConfig":{"$ref":"#/components/schemas/FieldMaskingConfig"},"allowedMethods":{"type":"string"},"status":{"type":"string","enum":["DRAFT","ACTIVE","INACTIVE"]},"tenantId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"public":{"type":"boolean"},"idValue":{"type":"string"},"fieldConfig":{"type":"string"}}},"DataEndpointId":{"type":"object","properties":{"id":{"type":"string","format":"uuid"}}},"DomainEvent":{"type":"object","properties":{"eventType":{"type":"string"},"eventId":{"type":"string","format":"uuid"},"occurredOn":{"type":"string","format":"date-time"}}},"FieldMaskingConfig":{"type":"object","properties":{"configJson":{"type":"string"},"empty":{"type":"boolean"}}},"TableMetadata":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"description":{"type":"string"},"visibility":{"type":"string","enum":["VISIBLE","HIDDEN","TECHNICAL","EVERYWHERE","DETAIL_ONLY"]},"lastSyncedAt":{"type":"string","format":"date-time"},"columns":{"type":"array","items":{"$ref":"#/components/schemas/ColumnMetadata"}}}},"UpdateConnectorRequest":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"config":{"type":"object","additionalProperties":{"type":"object"}},"registeredTables":{"type":"array","items":{"$ref":"#/components/schemas/TableMetadata"}},"active":{"type":"boolean"}}},"ApiKeyResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"tenantId":{"type":"string"},"name":{"type":"string"},"prefix":{"type":"string"},"status":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"lastUsedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"CreateTenantRequest":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}}},"CreateApiKeyRequest":{"type":"object","properties":{"name":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"}}},"ApiKeyCreationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"tenantId":{"type":"string"},"name":{"type":"string"},"rawKey":{"type":"string"},"prefix":{"type":"string"},"status":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"CreateConnectorRequest":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"config":{"type":"object","additionalProperties":{"type":"object"}}}},"ApiKeyListItemResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"prefix":{"type":"string"},"status":{"type":"string"},"lastUsedAt":{"type":"string","format":"date-time"},"expiresAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}}}}}